move ldapUtils from helpers to utils
